refactor: structure refactoring
