--- src/app/core/core.component.html
@@ -1,1 +1,0 @@
-<p>core works!</p>
