--- src/app/core/core.component.ts
@@ -1,7 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-core',
-  templateUrl: './core.component.html',
-})
-export class CoreComponent {}
