+++ src/app/main/main-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { MainPageComponent } from './pages/main-page/main-page.component';
+import { MainPageComponent } from './pages/main-page.component';
 
 const routes: Routes = [{ path: '', component: MainPageComponent }];
