--- src/app/shared/shared.component.html
@@ -1,1 +1,0 @@
-<p>shared works!</p>
