--- src/app/shared/shared.component.ts
@@ -1,7 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-shared',
-  templateUrl: './shared.component.html',
-})
-export class SharedComponent {}
