+++ src/app/shopping-cart/shopping-cart-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { CartPageComponent } from './pages/cart-page/cart-page.component';
+import { CartPageComponent } from './pages/cart-page.component';
 
 const routes: Routes = [{ path: '', component: CartPageComponent }];
