+++ src/app/shopping-cart/shopping-cart.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { ShoppingCartRoutingModule } from './shopping-cart-routing.module';
-import { CartPageComponent } from './pages/cart-page/cart-page.component';
+import { CartPageComponent } from './pages/cart-page.component';
 
 @NgModule({
   declarations: [CartPageComponent],
